% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macamseq-package.R
\docType{package}
\name{macamseq-package}
\alias{macamseq}
\alias{macamseq-package}
\title{macamseq: A subset of the R package macam for DNA sequence analysis}
\description{
macamesq contains a subset of the R functions implemented in macam. The functions are mainly for DNA sequence analysis.
}
\keyword{internal}
